Move WhoAmI controller under v1 package.
