+++ st2api/st2api/controllers/root.py
@@ -13,15 +13,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from oslo_config import cfg
-
 from st2common import __version__
-from st2common.rbac import utils as rbac_utils
-from st2common.services.rbac import get_roles_for_user
 
 __all__ = [
-    'RootController',
-    'WhoAmIController'
+    'RootController'
 ]
 
 
@@ -40,37 +35,4 @@ class RootController(object):
         return data
 
 
-class WhoAmIController(object):
-    def get(self, requester_user, auth_info):
-        """
-        Meta API endpoint wich returns information about the currently authenticated user.
-
-            Handle:
-                GET /v1/whoami
-        """
-
-        data = {}
-
-        if cfg.CONF.rbac.enable and requester_user:
-            role_dbs = get_roles_for_user(user_db=requester_user)
-            roles = [role_db.name for role_db in role_dbs]
-        else:
-            roles = []
-
-        data = {
-            'username': requester_user.name,
-            'authentication': {
-                'method': auth_info['method'],
-                'location': auth_info['location']
-            },
-            'rbac': {
-                'enabled': cfg.CONF.rbac.enable,
-                'roles': roles,
-                'is_admin': rbac_utils.user_is_admin(user_db=requester_user)
-            },
-        }
-        return data
-
-
 root_controller = RootController()
-whoami_controller = WhoAmIController()
